# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "Sandbox Connect API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/admin/outputs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "List outputs pending NHA approval",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum outputs (1-200)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.PendingOutputsResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error403"
                        }
                    }
                }
            }
        },
        "/v1/admin/outputs/{output_id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Queues idempotent Files Connect publication into the owner's object-backed workspace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "Approve and publish output files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique request key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ApprovalResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/outputs/{output_id}/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "List CSV files for an output awaiting approval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.OutputFilesResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/outputs/{output_id}/files/{file_id}/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "output-admin"
                ],
                "summary": "Preview a pending CSV output file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Stable file ID",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sandbox-backend-service_internal_filesconnect.Preview"
                        }
                    }
                }
            }
        },
...
        "/v1/notebooks/{notebook_name}/outputs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stops an owned, ready notebook and queues an asynchronous output run. In booking mode the linked booking must be active.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outputs"
                ],
                "summary": "Submit a notebook output run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notebook Name",
                        "name": "notebook_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique request key",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.OutputResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error409"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error500"
                        }
                    }
                }
            }
        },
        "/v1/outputs/{output_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an owned output run's status without exposing review-bucket keys.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outputs"
                ],
                "summary": "Get an output run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Output ID",
                        "name": "output_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.OutputResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Error404"
                        }
                    }
                }
            }
        },
...
        "/v1/workspace/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "List approved output files in the user's workspace",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.WorkspaceFilesResponse"
                        }
                    }
                }
            }
        },
        "/v1/workspace/files/{file_id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Get a short-lived download URL for an approved workspace CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable file ID",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sandbox-backend-service_internal_filesconnect.Download"
                        }
                    }
                }
            }
        },
        "/v1/workspace/files/{file_id}/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Preview an approved workspace CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable file ID",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sandbox-backend-service_internal_filesconnect.Preview"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cmd_api.ApprovalResponse": {
            "type": "object",
            "properties": {
                "outputId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "cmd_api.Error403": {
            "description": "Response for permission errors",
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Error message",
                    "type": "string",
                    "example": "string"
                },
                "title": {
                    "description": "Error title",
                    "type": "string",
                    "example": "Forbidden"
                },
                "type": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "cmd_api.OutputFilesResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd_api.OutputFile"
                    }
                },
                "outputId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "cmd_api.PendingOutputsResponse": {
            "type": "object",
            "properties": {
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd_api.OutputResponse"
                    }
                }
            }
        },
...
        "sandbox-backend-service_internal_filesconnect.Download": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}